+++ pysys-examples/internal/testcases/PySys_internal_133/run.py
@@ -1,0 +1,14 @@
+import pysys
+from pysys.constants import *
+
+class PySysTest(pysys.basetest.BaseTest):
+
+	def execute(self):
+		self.assertThat('actual == expected', actual=pysys.utils.fileutils.loadJSON(self.input+'/test.json'), expected={
+			'PoundSign':['\xa3'],
+			'Number': [123, 456.789]
+		})
+
+		
+	def validate(self):
+		pass